+++ src/main/java/com/octo/controller/filter/AuthenticationFilter.java
@@ -4,14 +4,17 @@ import java.io.IOException;
 import java.lang.reflect.Method;
 
 import javax.annotation.security.PermitAll;
+import javax.annotation.security.RolesAllowed;
 import javax.ws.rs.container.ContainerRequestContext;
 import javax.ws.rs.container.ContainerRequestFilter;
 import javax.ws.rs.container.ResourceInfo;
 import javax.ws.rs.core.Context;
 import javax.ws.rs.ext.Provider;
 
+import org.apache.commons.lang3.ArrayUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 
+import com.octo.model.authentication.UserRoleType;
 import com.octo.model.entity.User;
 import com.octo.model.error.ErrorType;
 import com.octo.model.error.GlobalException;
@@ -40,6 +43,11 @@ public class AuthenticationFilter implements ContainerRequestFilter {
     @Autowired
     private UserService userService;
 
+    /**
+     * Default authentication error.
+     */
+    private GlobalException authenticationError = new GlobalException(ErrorType.AUTHORIZATION_FAILED, "authentication");
+
     @Override
     public final void filter(final ContainerRequestContext requestContext) throws IOException {
         Method method = resourceInfo.getResourceMethod();
@@ -48,11 +56,55 @@ public class AuthenticationFilter implements ContainerRequestFilter {
             return;
         }
 
+        if (!method.isAnnotationPresent(RolesAllowed.class)) {
+            throw new GlobalException(ErrorType.INTERNAL_ERROR, "Missing RolesAllowed annotation.");
+        }
+
         User user = new UserMapper().apply(requestContext);
 
-        if (!Constants.DEFAULT_ADMIN_LOGIN.equals(user.getLogin())
-                || !userService.isDefaultAdminitratorAllowed(user.getPassword())) {
-            throw new GlobalException(ErrorType.AUTHORIZATION_FAILED, "authentication");
+        RolesAllowed annotation = method.getAnnotation(RolesAllowed.class);
+        String[] roles = annotation.value();
+
+        if (Constants.AUTHENTICATION_BASIC_SCHEME.equals(user.getAuthenticationType())) {
+            this.validateBasicAuthentitacion(roles, user);
+        } else {
+            this.validateTokenAuthentitacion(roles, user);
+        }
+    }
+
+    /**
+     * Validate basic authentication.
+     *
+     * @param roles
+     *            Valid roles.
+     * @param user
+     *            User to validate.
+     */
+    public void validateBasicAuthentitacion(final String[] roles, final User user) {
+        if (!Constants.DEFAULT_ADMIN_LOGIN.equals(user.getLogin())) {
+            throw authenticationError;
+        }
+        if (!userService.isDefaultAdminitratorAllowed(user.getPassword())) {
+            throw authenticationError;
+        }
+    }
+
+    /**
+     * Validate token authentication.
+     *
+     * @param roles
+     *            Valid roles.
+     * @param userToken
+     *            User with token.
+     */
+    public void validateTokenAuthentitacion(final String[] roles, final User userToken) {
+        if (!ArrayUtils.contains(roles, UserRoleType.TOKEN)) {
+            throw new GlobalException(ErrorType.AUTHORIZATION_ERROR, "Token are not allowed.");
+        }
+
+        User user = userService.getUserFromToken(userToken.getPassword());
+        if (user == null || !Constants.DEFAULT_ADMIN_LOGIN.equals(user.getLogin())) {
+            throw authenticationError;
         }
     }
 }