+++ src/main/java/com/octo/model/dto/user/token/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * All user token DTO.
+ *
+ * @author Vincent Moittié
+ */
+package com.octo.model.dto.user.token;