+++ src/main/java/com/octo/service/UserService.java
@@ -1,5 +1,7 @@
 package com.octo.service;
 
+import java.security.NoSuchAlgorithmException;
+import java.security.SecureRandom;
 import java.util.Base64;
 import java.util.Collections;
 import java.util.List;
@@ -7,6 +9,7 @@ import java.util.Optional;
 
 import javax.transaction.Transactional;
 
+import org.apache.commons.lang3.RandomStringUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.validator.routines.EmailValidator;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,7 +20,9 @@ import com.octo.dao.IDAO;
 import com.octo.model.authentication.UserRoleType;
 import com.octo.model.dto.user.SearchUserDTO;
 import com.octo.model.dto.user.UserViewDTO;
+import com.octo.model.dto.user.token.SearchUserTokenDTO;
 import com.octo.model.entity.User;
+import com.octo.model.entity.UserToken;
 import com.octo.model.error.ErrorType;
 import com.octo.model.error.GlobalException;
 import com.octo.utils.Constants;
@@ -36,6 +41,12 @@ public class UserService {
      */
     @Autowired
     private IDAO<User, QueryFilter> userDAO;
+
+    /**
+     * User DAO.
+     */
+    @Autowired
+    private IDAO<UserToken, QueryFilter> userTokenDAO;
 
     /**
      * Password encoder.
@@ -143,5 +154,127 @@ public class UserService {
             return List.of(UserRoleType.ADMIN);
         }
         return Collections.emptyList();
+    }
+
+    /**
+     * Get user from token.
+     *
+     * @param token
+     *            Token to find user.
+     * @return User or null.
+     */
+    public User getUserFromToken(final String token) {
+        Optional<UserToken> userToken = this.userTokenDAO.load(new SearchUserTokenDTO(token));
+        if (userToken.isEmpty()) {
+            return null;
+        }
+        return userToken.get().getUser();
+    }
+
+    /**
+     * Get user tokens.
+     *
+     * @param login
+     *            User login.
+     * @return Tokens.
+     */
+    public List<String> getUserToken(final String login) {
+        SearchUserDTO userFilter = new SearchUserDTO();
+        userFilter.setLogin(login);
+        Optional<User> user = userDAO.load(userFilter);
+        if (user.isEmpty()) {
+            throw new GlobalException(ErrorType.ENTITY_NOT_FOUND, "user", login);
+        }
+
+        SearchUserTokenDTO tokenFilter = new SearchUserTokenDTO();
+        tokenFilter.setUser(user.get().getId().toString());
+
+        return userTokenDAO.find(tokenFilter).stream().map(UserToken::getName).toList();
+    }
+
+    /**
+     * Create token for user.
+     *
+     * @param login
+     *            User login.
+     * @param name
+     *            Token name.
+     * @return Generated token.
+     * @throws NoSuchAlgorithmException
+     *             On no secure random algorithm found.
+     */
+    public String createToken(final String login, final String name) throws NoSuchAlgorithmException {
+        SearchUserDTO userFilter = new SearchUserDTO();
+        userFilter.setLogin(login);
+        Optional<User> user = userDAO.load(userFilter);
+        if (user.isEmpty()) {
+            throw new GlobalException(ErrorType.ENTITY_NOT_FOUND, "user", login);
+        }
+
+        List<String> tokens = this.getUserToken(login);
+        String tokenName = StringUtils.upperCase(name);
+        if (tokens.contains(tokenName)) {
+            throw new GlobalException(ErrorType.WRONG_VALUE, "name", "duplicate");
+        }
+
+        String token = this.generateToken();
+        UserToken userToken = new UserToken();
+        userToken.setUser(user.get());
+        userToken.setName(tokenName);
+        userToken.setToken(passwordEncoder.encode(token));
+
+        this.userTokenDAO.save(userToken);
+
+        return token;
+    }
+
+    /**
+     * Generate token.
+     *
+     * @return Token.
+     * @throws NoSuchAlgorithmException
+     *             On no secure random algorithm found.
+     */
+    public String generateToken() throws NoSuchAlgorithmException {
+        String token = RandomStringUtils.random(Constants.TOKEN_LENGHT, 0, 0, true, true, null,
+                SecureRandom.getInstanceStrong());
+        User user = this.getUserFromToken(token);
+
+        if (user == null) {
+            return token;
+        }
+        return this.generateToken();
+    }
+
+    /**
+     * Delete user token.
+     *
+     * @param login
+     *            User login.
+     * @param filter
+     *            Filter to find token name.
+     */
+    public void deleteToken(final String login, final SearchUserTokenDTO filter) {
+        if (StringUtils.isBlank(filter.getName())) {
+            throw new GlobalException(ErrorType.EMPTY_VALUE, "name");
+        }
+        SearchUserDTO userFilter = new SearchUserDTO();
+        userFilter.setLogin(login);
+        Optional<User> user = userDAO.load(userFilter);
+        if (user.isEmpty()) {
+            throw new GlobalException(ErrorType.ENTITY_NOT_FOUND, "user", login);
+        }
+
+        SearchUserTokenDTO tokenFilter = new SearchUserTokenDTO();
+        tokenFilter.setUser(user.get().getId().toString());
+        tokenFilter.setName(filter.getName());
+
+        Optional<UserToken> token = userTokenDAO.load(tokenFilter);
+
+        if (token.isEmpty()) {
+            throw new GlobalException(ErrorType.ENTITY_NOT_FOUND, "user token");
+        }
+
+        userTokenDAO.delete(token.get());
     }
 }