+++ src/test/java/com/octo/service/GroupServiceTest.java
@@ -1,7 +1,10 @@
 package com.octo.service;
 
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
+
+import java.util.Optional;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -16,6 +19,8 @@ import com.octo.dao.IDAO;
 import com.octo.model.entity.Group;
 import com.octo.model.entity.Project;
 import com.octo.model.entity.ProjectGroup;
+import com.octo.model.error.ErrorType;
+import com.octo.model.error.GlobalException;
 import com.octo.utils.predicate.filter.QueryFilter;
 
 @ExtendWith(SpringExtension.class)
@@ -44,15 +49,23 @@ class GroupServiceTest {
     }
 
     @Test
-    void testAddProjectToGroup() {
-        Group group = new Group();
-        group.setId(1L);
+    void testAddProjectToGroupWithUnknowGroup() {
+        Mockito.when(groupDAO.load(Mockito.any())).thenReturn(Optional.empty());
+        GlobalException exception = null;
+        try {
+            this.service.addProjectToGroup(1L, null);
+        } catch (GlobalException e) {
+            exception = e;
+        }
+        assertNotNull(exception);
+        assertEquals(exception.getError().getMessage(), ErrorType.ENTITY_NOT_FOUND.getMessage());
+    }
 
-        Mockito.when(groupDAO.loadEntityById(Mockito.any(), Mockito.any())).thenReturn(group);
+    @Test
+    void testAddProjectToGroup() {
+        Mockito.when(groupDAO.load(Mockito.any())).thenReturn(Optional.of(new Group()));
         Mockito.when(projectGroupDAO.save(Mockito.any())).thenReturn(null);
 
-        Project masterProject = new Project();
-        masterProject.setId(1L);
-        assertNull(service.addProjectToGroup(masterProject, new Project()));
+        assertNull(service.addProjectToGroup(1L, new Project()));
     }
 }